refactor(extension): move `workspace` into own class
