+++ src/constants.js
@@ -1,0 +1,17 @@
+export default {
+  PANEL_BOX: {
+    0: "_leftBox",
+    1: "_centerBox",
+    2: "_rightBox"
+  },
+  OFF: 0,
+  REDUCE_OPACITY: 1,
+  DESATURATE: 2,
+  NO_LIMIT: 100,
+  LOW_OPACITY: 175,
+  ICONS_SIZE: 10,
+  TEXTURES_SIZE: 20,
+  LEFT_CLICK: 1,
+  MIDDLE_CLICK: 2,
+  RIGHT_CLICK: 3
+}
